feat: load flow from the server on edit page
